Reorganize jobs, use Sidekiq jobs to prep for future refactor
